Добавить обработку ошибки декодирования

+++ engine/injection/controllers/data_file_controller.go
@@ -57,8 +57,12 @@ func (c *DataFileController) DeleteDataFile(
 	request *http.Request) {
 
 	var deleteIntent filters.IdsFilter
-	Decode(request, &deleteIntent, writer)
-
+	decodeError := Decode(request, &deleteIntent)
+	if decodeError != nil {
+		writer.WriteHeader(http.StatusBadRequest)
+		writer.Write([]byte(decodeError.Error()))
+		return
+	}
 	str := strings.Trim(strings.Replace(fmt.Sprint(deleteIntent.Ids), " ", ",", -1), "[]")
 	query := "delete from project_config_data where data_id in ($1)"
 	query = strings.Replace(query, "$1", str, -1)
